+++ docker/test/start.sh
@@ -50,6 +50,8 @@ err(){
 # Turn on debug in containers (restconf, backend)
 DBG=${DBG:-0}
 
+NAME=${NAME:-"clixon-test"}
+
 # Expose other host port than port 80
 PORT=${PORT:-8080}
 
@@ -85,7 +87,7 @@ CONFIG=${CONFIG:-$CONFIG0}
 
 # Create clixon-test container
 >&2 echo -n "Starting Container..."
-sudo docker run -p $PORT:80 -p $SPORT:443 --name clixon-test --rm -e DBG=$DBG -e CONFIG="$CONFIG" -e STORE="$STORE" -td clixon/clixon-test || err "Error starting clixon-test"
+sudo docker run -p $PORT:80 -p $SPORT:443 --name ${NAME} --rm -e DBG=$DBG -e CONFIG="$CONFIG" -e STORE="$STORE" -td clixon/clixon-test || err "Error starting clixon-test"
 
 # Wait for snmpd to start
 sudo docker exec -t clixon-test bash -c 'while [ ! -S /var/run/snmp.sock ]; do sleep 1; done'